move time interpolation out of nudging.py to  interpolate.py
