+++ workflows/prognostic_c48_run/runtime/nudging.py
@@ -19,6 +19,7 @@ from typing import (
     Optional,
 )
 import logging
+from runtime.interpolate import time_interpolate_func
 
 logger = logging.getLogger(__name__)
 
@@ -94,7 +95,7 @@ def setup_get_reference_state(
 
     initial_time_label = config.reference_initial_time
     if initial_time_label is not None:
-        get_reference_state = _time_interpolate_func(
+        get_reference_state = time_interpolate_func(
             get_reference_state,
             initial_time=_label_to_time(initial_time_label),
             frequency=timedelta(seconds=config.reference_frequency_seconds),
@@ -163,47 +164,6 @@ def _label_to_time(time: str) -> cftime.DatetimeJulian:
         int(time[13:15]),
     )
 
-
-def _time_interpolate_func(
-    func: Callable[[cftime.DatetimeJulian], dict],
-    frequency: timedelta,
-    initial_time: cftime.DatetimeJulian,
-) -> Callable[[cftime.DatetimeJulian], dict]:
-    cached_func = functools.lru_cache(maxsize=2)(func)
-
-    @functools.wraps(cached_func)
-    def myfunc(time: cftime.DatetimeJulian) -> State:
-        quotient = (time - initial_time) // frequency
-        remainder = (time - initial_time) % frequency
-
-        state: State = {}
-        if remainder == timedelta(0):
-            state.update(cached_func(time))
-        else:
-            begin_time = quotient * frequency + initial_time
-            end_time = begin_time + frequency
-
-            state_0 = cached_func(begin_time)
-            state_1 = cached_func(end_time)
-
-            state.update(
-                _average_states(state_0, state_1, weight=(end_time - time) / frequency)
-            )
-
-        return state
-
-    return myfunc
-
-
-def _average_states(state_0: State, state_1: State, weight: float) -> State:
-    common_keys = set(state_0) & set(state_1)
-    out = {}
-    for key in common_keys:
-        if isinstance(state_1[key], xr.DataArray):
-            out[key] = (
-                state_0[key] * weight + (1 - weight) * state_1[key]  # type: ignore
-            )
-    return out
 
 
 def get_nudging_tendency(
